#42 bodyCode 자료형 Long에서 int로 변경

+++ src/main/java/A1B1O3/bodyrecord/auth/domain/Token.java
@@ -17,7 +17,7 @@ import static lombok.AccessLevel.PROTECTED;
 public class Token {
 
     @Id
-    private Long memberCode;
+    private int memberCode;
 
     private String refreshToken;
 
@@ -25,7 +25,7 @@ public class Token {
         this.refreshToken = refreshToken;
     }
 
-    public static Token of(Long memberCode, String refreshToken) {
+    public static Token of(int memberCode, String refreshToken) {
         return new Token(memberCode, refreshToken);
     }
 }
