+++ src/main/java/A1B1O3/bodyrecord/member/dto/response/MemberDetailResponse.java
@@ -13,7 +13,7 @@ import static lombok.AccessLevel.PRIVATE;
 @Getter
 @RequiredArgsConstructor(access = PRIVATE)
 public class MemberDetailResponse {
-    private final Long memberCode;
+    private final int memberCode;
     private final String memberSocialid;
     private final String memberName;
     private final String memberNickname;
